requirements.txt
absl-py==2.1.0
academictorrents==2.3.3
accelerate==1.10.1
addict==2.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
banal==1.0.6
bencode.py==2.0.0
bitstring==3.1.5
blinker==1.9.0
braceexpand==0.1.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.2.1
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.3.1
cupy-cuda12x==13.4.0
cycler==0.12.1
databricks-cli==0.18.0
dataset==1.6.2
datasets==3.3.2
decorator==5.2.1
diffusers==0.32.2
dill==0.3.8
docker==6.1.3
docstring_parser==0.16
einops==0.8.1
entrypoints==0.4
executing==2.2.1
facexlib==0.3.0
fastrlock==0.8.3
filelock==3.17.0
filterpy==1.4.5
Flask==2.3.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==0.16.0
geomloss==0.2.6
geoopt==0.5.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.1.1
grpcio==1.70.0
gunicorn==20.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
icecream==2.1.7
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib-metadata==6.11.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonargparse==4.37.0
keopscore==2.3
kiwisolver==1.4.8
kmeans-pytorch==0.3
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.1
lightning-utilities==0.14.3
llvmlite==0.44.0
lmdb==1.7.3
loguru==0.7.3
lpips==0.1.4
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mlflow==2.4.2
modelscope==1.29.1
monai==1.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nibabel==5.3.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
nvitop==1.4.2
oauthlib==3.3.1
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==23.2
pandas==2.2.3
pillow==11.1.0
piq==0.8.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.3.0
protobuf==4.25.8
psutil==7.0.0
pyarrow==12.0.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyiqa==0.1.12
PyJWT==2.10.1
pyparsing==3.2.1
PyPubSub==3.3.0
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.1
pytorch-lightning==2.5.0.post0
pytorch-msssim==1.0.0
pytz==2023.4
PyYAML==6.0.2
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.10.1
sentencepiece==0.2.1
shapely==2.1.1
SimpleITK==2.4.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==1.4.54
sqlparse==0.5.3
stringzilla==3.12.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.19
tokenizers==0.22.1
tomli==2.2.1
torch==2.4.1
torchinfo==1.8.0
torchmetrics==1.6.2
torchvision==0.19.1
tqdm==4.67.1
transformers==4.57.1
triton==3.0.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webdataset==0.2.111
websocket-client==1.8.0
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
